% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentration.R
\name{calc_concCoef}
\alias{calc_concCoef}
\title{Calculates the Concentration Coefficient}
\usage{
calc_concCoef(x, y, w = NULL)
}
\arguments{
\item{x}{A vector of a income source}

\item{y}{The vector of total income (according to which the income source x will be sorted)}

\item{w}{(optional) A vector of sample weights}
}
\value{
Returns the numeric value of the Concentration Coefficient
}
\description{
Calculates the Concentration Coefficient
}
