% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gini.R
\name{calc_gini_fromCDF}
\alias{calc_gini_fromCDF}
\title{Calculates the Gini Coefficient from a CDF function}
\usage{
calc_gini_fromCDF(
  cdf,
  boundedIntegral = F,
  p_lowerBound_integration = 0.004,
  p_upperBound_integration = 0.997
)
}
\arguments{
\item{cdf}{A CDF function}

\item{boundedIntegral}{A logical value. Indicates if the numerical integral should be performed only over a given bounded/finite interval. In this case p_lowerBound_integration and p_upperBound_integration should be informed}

\item{p_lowerBound_integration}{Lower bound of the finite interval over which the integral will be calculated}

\item{p_upperBound_integration}{Upper bound of the finite interval over which the integral will be calculated}
}
\value{
Returns the numeric value of the Gini Coefficient
}
\description{
Calculates the Gini Coefficient from a CDF function
}
