% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gini.R
\name{calc_giniCoef}
\alias{calc_giniCoef}
\title{Calculates the Gini Coefficient}
\usage{
calc_giniCoef(x, w = NULL)
}
\arguments{
\item{x}{A vector of a income source}

\item{w}{(optional) A vector of sample weights}
}
\value{
Returns the numeric value of the Gini Coefficient
}
\description{
Calculates the Gini Coefficient
}
