% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gini.R
\name{calc_giniCoef_fromQuantile}
\alias{calc_giniCoef_fromQuantile}
\title{Calculates the Gini Coefficient from a quantile function}
\usage{
calc_giniCoef_fromQuantile(qf, gridIntegration = NULL)
}
\arguments{
\item{gridIntegration}{(optional) A grid of class 'NIGrid' for multivariate numerical integration (mvQuad package)}

\item{quantile_func}{A quantile function}
}
\value{
Returns the numeric value of the Gini Coefficient
}
\description{
Calculates the Gini Coefficient from a quantile function
}
