% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_growthConcCoef.R
\name{calc_growthConcCoef}
\alias{calc_growthConcCoef}
\title{Computes the Concentration Coefficient for the income growth between two periods}
\usage{
calc_growthConcCoef(x0, x1, w0 = NULL, w1 = NULL)
}
\arguments{
\item{x0}{A vector of incomes at time 0}

\item{x1}{A vector of incomes at time 1}

\item{w0}{(optional) A vector of sample weights at time 0}

\item{w1}{(optional) A vector of sample weights at time 1}
}
\value{
Returns the Growth Concentration Coenfficient (numeric value)
}
\description{
Computes the Concentration Coefficient for the income growth between two periods
}
