% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_growthConcCoef.R
\name{calc_growthConcCoef_fromQuantile}
\alias{calc_growthConcCoef_fromQuantile}
\title{Computes the Concentration Coefficient for the income growth between two periods using Quantile Functions as input}
\usage{
calc_growthConcCoef_fromQuantile(
  qf0,
  qf1,
  correct_for_negativeAppropriation = F,
  gridIntegration = NULL
)
}
\arguments{
\item{qf0}{A quantile function of the income distribution at time 0}

\item{qf1}{A quantile function of the income distribution at time 1}

\item{correct_for_negativeAppropriation}{A logical value. If \code{TRUE}, applies a alternative method that corrects for negative values in the Growth Concentration Curve. Default to \code{FALSE}.}

\item{gridIntegration}{(optional) A grid of class 'NIGrid' for multivariate numerical integration (mvQuad package)}
}
\value{
Returns the Growth Concentration Coefficient (numeric value)
}
\description{
Computes the Concentration Coefficient for the income growth between two periods using Quantile Functions as input
}
