% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_distributions.R
\name{combine_distributions}
\alias{combine_distributions}
\title{Combines two different distributions that were fitted using inequalityTools}
\usage{
combine_distributions(
  fit1,
  fit2,
  n1,
  n2,
  quadrature_precision = 5000,
  points_to_adjust = 2000,
  upper_limit = 1e+05,
  upper_p_limit = 0.9995
)
}
\arguments{
\item{fit1}{A interpolated distribution fitted with tableFit_nonParametric or tableFit_gb2}

\item{fit2}{Another interpolated distribution fitted with tableFit_nonParametric or tableFit_gb2}

\item{n1}{The number of observations or proportion of the fit1.}

\item{n2}{The number of observations or proportion of the fit2.}

\item{quadrature_precision}{Gauss-Legendre Quadrature level - argument to be passed to mvQuad::createNIGrid. Default to 5e3}

\item{points_to_adjust}{The number of points from which the cdf, quantile function and Lorenz Curve will be calculated. Default to 2000}

\item{upper_limit}{The maximum income value of the distribution support. Default to 1e5}

\item{upper_p_limit}{A cumulative probability value which will be taken as the maximum for the quantile function. After it, quantile function will be top coded. Default to .9995}
}
\value{
Returns a list with the following components
\itemize{
  \descriptive A tibble with two columns: estMean (the distribution mean), gini (the Gini Coefficient, if \code{gini = TRUE}, NA otherwise)
  \pdf The PDF function. Takes values as input and returns probability densities.
  \cdf The CDF function. Takes values as input and returns cumulative probabilities.
  \quantile The Quantile function. Takes cumulative probabilities as input and returns quantile values as outputs.
  \lorenz The Lorenz Curve function. Takes cumulative probabilities as input and returns the proportion of income accumulated at the quantile.
}
}
\description{
Combines two distributions fitted with inequalityTools by weighted summing their PDFs. Then recalculates the other distribution functions (pdf, cdf, quantile function, and Lorenz Curve) and descriptives.
}
