% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_giniBySources.R
\name{decomp_giniBySources}
\alias{decomp_giniBySources}
\title{Static decomposition the Gini Coefficient by Income Sources}
\usage{
decomp_giniBySources(data, incomeSources_vars, totalIncome_var, weight = NULL)
}
\arguments{
\item{data}{A data.frame with income sources and the total income (and possibliy sample weights) as variables}

\item{incomeSources_vars}{A character vector with names of the income sources (they must be columns of 'data')}

\item{totalIncome_var}{A character value naming the variable wich represents the total income (it must be a column of 'data' and for each observation in the data, the income sources must add up to the total income).}

\item{weight}{(optional) A character values naming the variable wich represents the sample weights}
}
\value{
Returns a tibble with the following coluns:
    - Income Source (name of the income source)
    - Concentration Coefficient (with respect to the total income)
    - Share (with respect to the total income)
    - Concentration Coefficient*Share (sums up to the Gini Coefficient)
    - Contribution to Gini Coefficient
}
\description{
Static decomposition the Gini Coefficient by Income Sources
}
