% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_meanIncomeBySources.R
\name{decomp_meanIncomeBySources}
\alias{decomp_meanIncomeBySources}
\title{Static decomposition of the mean income by sources}
\usage{
decomp_meanIncomeBySources(
  data,
  incomeSources_vars,
  totalIncome_var,
  weight = NULL
)
}
\arguments{
\item{data}{A data.frame with income sources and the total income (and possibliy sample weights) as variables}

\item{incomeSources_vars}{A character vector with names of the income sources (they must be columns of 'data')}

\item{totalIncome_var}{A character value naming the variable wich represents the total income (it must be a column of 'data' and for each observation in the data, the income sources must add up to the total income).}

\item{weight}{(optional) A character values naming the variable wich represents the sample weights}
}
\value{
Returns a tibble with the following coluns:
    - Income Source: name of the income source
    - Mean: Mean of the income source
    - Prop. Pop. Recipients: Proportion of the population receiving non-zero income for each source
    - Mean*Prop. Pop. Recipients: mean times the proportion of recipients. This column must sum up to the global mean of the total income variable.
}
\description{
Static decomposition of the mean income by sources
}
