% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_giniBySources.R
\name{dynamicDecomp_giniBySources}
\alias{dynamicDecomp_giniBySources}
\title{Dynamic decomposition of the Gini Coefficient by Income Sources}
\usage{
dynamicDecomp_giniBySources(decomp_giniBySources0, decomp_giniBySources1)
}
\arguments{
\item{decomp_giniBySources0}{A data.frame with with results of a static decomposition the Gini (produced by decomp_giniBySources) for group/time 0}

\item{decomp_giniBySources1}{A data.frame with with results of a static decomposition the Gini (produced by decomp_giniBySources) for group/time 1}
}
\value{
Returns a tibble with the following coluns:
    - Income Source (name of the income source)
    - Concentration Coefficient (with respect to the total income)
    - Share (with respect to the total income)
    - Concentration Coefficient*Share (sums up to the Gini Coefficient)
    - Contribution to Gini Coefficient
}
\description{
Dynamic decomposition of the Gini Coefficient by Income Sources
}
