% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthAppr_byStrata.R
\name{growthAppr_byStrata}
\alias{growthAppr_byStrata}
\title{Computes the Growth Appropriation between two periods for equal-sized income strata/bins}
\usage{
growthAppr_byStrata(x0, x1, n_strata = 10, w0 = NULL, w1 = NULL)
}
\arguments{
\item{x0}{A vector of incomes at time 0}

\item{x1}{A vector of incomes at time 1}

\item{n_strata}{A numeric value indicating how many bins will be created}

\item{w0}{(optional) A vector of sample weights at time 0}

\item{w1}{(optional) A vector of sample weights at time 1}
}
\value{
Returns a tibble with the following columns:
    - Strata: Number of the ordered income strata
    - Pop. Share: Population share at the strata
    - Mean 0: Mean income of the strata at time 0
    - Mean 1: Mean income of the strata at time 1
    - Growth Appropriation: Growth appropriation of the strata between times 0 and 1
}
\description{
Computes the Growth Appropriation between two periods for equal-sized income strata/bins
}
