% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cdf.R
\name{make_cdf}
\alias{make_cdf}
\title{Makes a Cumulative Distribution Function (CDF) out of a income vector}
\usage{
make_cdf(x, w = NULL)
}
\arguments{
\item{x}{A vector of a income source}

\item{w}{(optional) A vector of sample weights}
}
\value{
Returns a function which takes a vector of incomes as inputs (p) and gives points at the CDF as outputs
}
\description{
Makes a Cumulative Distribution Function (CDF) out of a income vector
}
