% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cdf.R
\name{make_cdf_fromQuantile}
\alias{make_cdf_fromQuantile}
\title{Makes a Cumulative Distribution Function (CDF) out of a quantile function}
\usage{
make_cdf_fromQuantile(qf)
}
\arguments{
\item{qf}{A quantile function}
}
\value{
Returns a function which takes a vector of incomes as inputs (p) and gives points at the CDF as outputs
}
\description{
Makes a Cumulative Distribution Function (CDF) out of a quantile function
}
