% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentration.R
\name{make_concentrationCurve}
\alias{make_concentrationCurve}
\title{Makes a Concentration Curve Function}
\usage{
make_concentrationCurve(x, y, w = NULL)
}
\arguments{
\item{x}{A vector of a income source}

\item{y}{The vector of total income (according to which the income source x will be sorted)}

\item{w}{(optional) A vector of sample weights}
}
\value{
Returns a function which takes a vector of probabilities as inputs (p) and gives points at the Concentration Curve as outputs
}
\description{
Makes a Concentration Curve Function
}
