% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genLorenz.R
\name{make_genLorenz}
\alias{make_genLorenz}
\title{Makes a Generalized Lorenz Curve Function out of a income vector}
\usage{
make_genLorenz(x, w = NULL)
}
\arguments{
\item{x}{A vector of incomes}

\item{w}{(optional) A vector of sample weights}
}
\value{
Returns a function which takes a vector of probabilities as inputs (p) and gives points at the Generalized Lorenz Curve as outputs
}
\description{
Makes a Generalized Lorenz Curve Function out of a income vector
}
