% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genLorenz.R
\name{make_genLorenz_fromQuantile}
\alias{make_genLorenz_fromQuantile}
\title{Makes a Generalized Lorenz Curve Function out of a quantile function}
\usage{
make_genLorenz_fromQuantile(quantile_func, gridIntegration = NULL)
}
\arguments{
\item{quantile_func}{A quantile function}

\item{gridIntegration}{(optional) A grid of class 'NIGrid' for multivariate numerical integration (mvQuad package)}
}
\value{
Returns a function which takes a vector of probabilities as inputs (p) and gives points at the Generalized Lorenz Curve as outputs
}
\description{
Makes a Generalized Lorenz Curve Function out of a quantile function
}
