% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthConcCurve.R
\name{make_growthConcCurve}
\alias{make_growthConcCurve}
\title{Makes a Growth Concentration Curve Function out of two income vectors}
\usage{
make_growthConcCurve(x0, x1, w0 = NULL, w1 = NULL, gridIntegration = NULL)
}
\arguments{
\item{x0}{A vector of incomes for group/time 0}

\item{x1}{A vector of incomes for group/time 1}

\item{w0}{(optional) A vector of sample weights for group/time 0}

\item{w1}{(optional) A vector of sample weights for group/time 1}

\item{gridIntegration}{(optional) A grid of class 'NIGrid' for multivariate numerical integration (mvQuad package)}
}
\value{
Returns a function which takes a vector of probabilities as inputs (p) and gives points at the Growth Concentration Curve as outputs
}
\description{
Makes a Growth Concentration Curve Function out of two income vectors
}
