% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthConcCurve.R
\name{make_growthConcCurve_fromQuantile}
\alias{make_growthConcCurve_fromQuantile}
\title{Makes a Growth Concentration Curve Function out of two quantile functions}
\usage{
make_growthConcCurve_fromQuantile(qf0, qf1, gridIntegration = NULL)
}
\arguments{
\item{qf0}{A quantile function for group/time 0}

\item{qf1}{A quantile function for group/time 1}

\item{gridIntegration}{(optional) A grid of class 'NIGrid' for multivariate numerical integration (mvQuad package)}
}
\value{
Returns a function which takes a vector of probabilities as inputs (p) and gives points at the Growth Concentration Curve as outputs
}
\description{
Makes a Growth Concentration Curve Function out of two quantile functions
}
