% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz.R
\name{make_lorenz_fromMeansByStrata}
\alias{make_lorenz_fromMeansByStrata}
\title{Makes a Lorenz Curve Function using information about means and size of income strata}
\usage{
make_lorenz_fromMeansByStrata(
  group_size,
  group_means,
  combine_function = "mean",
  precision = 5e-05
)
}
\arguments{
\item{group_size}{A numerical vector informing the size (i.e population of proportion) of each group}

\item{group_means}{A numerical vector informing the mean income of each group. Must have the same number of elements tha 'group_size'}
}
\value{
Returns a function which takes a vector of probabilities as inputs (p) and gives points at the Lorenz Curve as outputs
}
\description{
Makes a Lorenz Curve Function using information about means and size of income strata
}
