% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz.R
\name{make_lorenz_fromQuantile}
\alias{make_lorenz_fromQuantile}
\title{Makes a Lorenz Curve Function out of a quantile function}
\usage{
make_lorenz_fromQuantile(qf, gridIntegration = NULL)
}
\arguments{
\item{gridIntegration}{(optional) A grid of class 'NIGrid' for multivariate numerical integration (mvQuad package)}

\item{quantile_func}{A quantile function}
}
\value{
Returns a function which takes a vector of probabilities as inputs (p) and gives points at the Lorenz Curve as outputs
}
\description{
Makes a Lorenz Curve Function out of a quantile function
}
