% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeIncomeStrata.R
\name{makeIncomeStrata}
\alias{makeIncomeStrata}
\title{Cuts a continuous income variable into equal-sized bins}
\usage{
makeIncomeStrata(x, n_strata = 10, w = NULL)
}
\arguments{
\item{x}{A vector of incomes}

\item{n_strata}{A numeric value indicating how many bins will be created}

\item{w}{(optional) A vector of sample weights}
}
\value{
Returns a numeric vector with indicating which income strata each observation belongs to.
}
\description{
Cuts a continuous income variable into equal-sized bins
}
