% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanBystrata.R
\name{meanBystrata}
\alias{meanBystrata}
\title{Computes the mean income for a strata defined by a lower bound and an upper bound in the ordered cumulative income distribution}
\usage{
meanBystrata(x, lowerBound = 0, upperBound = 1, w = NULL)
}
\arguments{
\item{x}{A vector of incomes}

\item{lowerBound}{A value between 0 and 1, indicating the lower bound of the strata}

\item{upperBound}{A value between 0 and 1, indicating the upper bound of the strata}

\item{w}{(optional) A vector of sample weights}
}
\value{
Returns the mean (a numeric value) within the strata
}
\description{
Computes the mean income for a strata defined by a lower bound and an upper bound in the ordered cumulative income distribution
}
