% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanByStrata.R
\name{meanByStrata_fromQuantile}
\alias{meanByStrata_fromQuantile}
\title{Computes the mean income for a strata defined by a lower bound and an upper bound in the ordered cumulative income distribution from a Quantile Function}
\usage{
meanByStrata_fromQuantile(
  qf,
  lowerBound = 0,
  upperBound = 1,
  subdivisions = 2000
)
}
\arguments{
\item{qf}{A quantile function}

\item{lowerBound}{A value between 0 and 1, indicating the lower bound of the strata}

\item{upperBound}{A value between 0 and 1, indicating the upper bound of the strata}
}
\value{
Returns the mean (a numeric value) within the strata
}
\description{
Computes the mean income for a strata defined by a lower bound and an upper bound in the ordered cumulative income distribution from a Quantile Function
}
