% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poverty_fgt.R
\name{poverty_fgt}
\alias{poverty_fgt}
\title{Calculates the Foster-Greer-Thorbecke (FGT) family of poverty measures}
\usage{
poverty_fgt(x, z, w = NULL, alpha = 0)
}
\arguments{
\item{x}{A vector of a income source}

\item{z}{A numeric income value representing the poverty line}

\item{w}{(optional) A vector of sample weights}

\item{alpha}{An arbitrary inequality aversion parameter: alpha = 0 returns the poverty rate; alpha = 1 returns the average income gap. Default to 0.}
}
\value{
Returns the numeric value of the FGT measure of poverty
}
\description{
Calculates the Foster-Greer-Thorbecke (FGT) family of poverty measures
}
