% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poverty_senIndex.R
\name{poverty_senIndex}
\alias{poverty_senIndex}
\title{Calculates Amartya Sen's Index of poverty}
\usage{
poverty_senIndex(x, z, w = NULL)
}
\arguments{
\item{x}{A vector of a income source}

\item{z}{A numeric income value representing the poverty line}

\item{w}{(optional) A vector of sample weights}

\item{alpha}{An arbitrary inequality aversion parameter: alpha = 0 returns the poverty rate; alpha = 1 returns the average income gap. Default to 0.}
}
\value{
Returns the numeric value of the FGT measure of poverty
}
\description{
Calculates Amartya Sen's Index of poverty
}
\references{
Sen, Amartya. (1976) "Poverty: An ordinal approach to measurement." 
    Econometrica: Journal of the Econometric Society: 219-231.
}
