% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shareByStrata.R
\name{shareByStrata}
\alias{shareByStrata}
\title{Computes the income share of a strata defined by a lower bound and an upper bound in the ordered cumulative income distribution}
\usage{
shareByStrata(x, lowerBound = 0, upperBound = 1, w = NULL)
}
\arguments{
\item{x}{A vector of incomes}

\item{lowerBound}{A value between 0 and 1, indicating the lower bound of the strata}

\item{upperBound}{A value between 0 and 1, indicating the upper bound of the strata}

\item{w}{(optional) A vector of sample weights}
}
\value{
Returns income share (a numeric value) of the strata
}
\description{
Computes the income share of a strata defined by a lower bound and an upper bound in the ordered cumulative income distribution
}
