% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shareByStrata.R
\name{shareByStrata_fromQuantile}
\alias{shareByStrata_fromQuantile}
\title{Using a quantile function, computes the income share of a strata defined by a lower bound and an upper bound in the ordered cumulative income distribution}
\usage{
shareByStrata_fromQuantile(
  qf,
  lowerBound = 0,
  upperBound = 1,
  subdivisions = 2000
)
}
\arguments{
\item{qf}{A quantile funcion}

\item{lowerBound}{A value between 0 and 1, indicating the lower bound of the strata}

\item{upperBound}{A value between 0 and 1, indicating the upper bound of the strata}
}
\value{
Returns income share (a numeric value) of the strata
}
\description{
Using a quantile function, computes the income share of a strata defined by a lower bound and an upper bound in the ordered cumulative income distribution
}
