% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_growthConcCoef.R
\name{calc_growthConcCoef}
\alias{calc_growthConcCoef}
\title{Computes the Concentration Coefficient for the income growth between two periods}
\usage{
calc_growthConcCoef(
  x0,
  x1,
  w0 = NULL,
  w1 = NULL,
  correct_for_negativeAppropriation = T,
  correct_for_negativeGrowth = T,
  gridIntegration = NULL
)
}
\arguments{
\item{x0}{A vector of incomes at time 0}

\item{x1}{A vector of incomes at time 1}

\item{w0}{(optional) A vector of sample weights at time 0}

\item{w1}{(optional) A vector of sample weights at time 1}

\item{correct_for_negativeAppropriation}{A logical value. Correct for negative values in the Growth Appropriation Curve. Default to TRUE.}

\item{correct_for_negativeGrowth}{A logical value. Correct for negative growth. Default to TRUE.}

\item{gridIntegration}{(optional) A grid of class 'NIGrid' for multivariate numerical integration (mvQuad package)}
}
\value{
Returns the Growth Concentration Coenfficient (numeric value)
}
\description{
Computes the Concentration Coefficient for the income growth between two periods
}
